requirements.txt
torch==2.0.1
torchvision==0.15.2
numpy==1.24.3
pandas==2.0.3
streamlit==1.24.0
pillow==9.5.0
tqdm==4.65.0